refactor(test): make HealthKit payload generator self-contained
